feat: 대시보드의 childtag를 다른 api를 이용하여 가져오도록 수정하고, 네트워크 연결이 안되어있어도, 일단 새로운 데이터를 가져오도록 수정
